+++ src/basic/view/index.js
